pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "tideline"
version = "0.1.0"
description = "Marine heatwave forecasting system"
requires-python = ">=3.11"
dependencies = [
    "xarray>=2024.1",
    "netcdf4>=1.6",
    "pandas>=2.2",
    "numpy>=1.26",
    "scipy>=1.12",
    "tqdm>=4.66",
    "scikit-learn>=1.4",
    "xgboost>=2.0",
    "boto3>=1.34",
    "google-cloud-storage>=2.16",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "marimo>=0.6",
    "shap>=0.45",
    "requests>=2.31",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "ruff>=0.4",
    "mypy>=1.9",
    "ipython",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
